backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
et-xmlfile==1.1.0
Faker==13.15.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.8.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
uuid==1.30
validate-docbr==1.10.0
